Link created Hostingprovider back to the originating ProviderRequest, add missing fields: Hostingprovider.description and evidence type
